[2단계 - 장바구니 기능] 디노(신종화) 미션 제출합니다. (#323)
* feat: price max 제한 예외처리

* refactor: 컨트롤러 메서드 리턴 타입 변경

* refactor: AutoWired 제거, 불필요한 console.log 제거

* feat: MethodArgumentNotValidException 이외의 예외 처리하는 메서드 추가

* refactor: 상품 삭제 요청 확인 테스트 로직 변경

* refactor: 스키마 조건 변경, 필요하지 않은 테스트 삭제

* refactor: Service 테스트 없애고 Dao 테스트로 변경

* docs: step 2 기능 요구사항 작성

* test: 사용자 전체 조회 테스트

* feat: 사용자 설정 페이지 연동

* feat: 사용자 전체 조회 기능 추가

* refactor: 사용자 전체 조회 테스트 로직 변경

* docs: 구현 완료한 기능 체크
설정 페이지 연동, 사용자 전체 조회 기능 구현

* feat: 장바구니 정보를 저장하기 위한 테이블 생성 쿼리 작성, 더미 데이터 삽입

* feat: 장바구니 페이지 연동 및 해당 사용자의 장바구니 물품 반환 기능 추가

* feat: 물건 장바구니에 담기 기능 추가

* refactor: 장바구니 POST 로직 PathVariable을 활용한 형태로 변경

* refactor: 장바구니 담은 물건 삭제 기능 추가

* docs: 기능 요구사항 변경

* refactor: Dao 인터페이스 제거

* refactor: 상품 수정, 삭제 기능에 중간 검증 로직 추가

* refactor: 불필요한 코드 제거

* test: dao 테스트 추가

* test: service 테스트 추가

* refactor: findMemberId 테스트 로직 변경

* refactor: getCartItem 메서드 이동

* test: pageController 테스트

* feat: view controller 적용

* feat: interceptor 적용

* feat: argumentResolver 적용

* refactor: 불필요한 구문 제거

* refactor: 상수 정의 대신 HttpHeaders에 정의된 String으로 변경

* refactor: extract 내부 로직 메서드 분리

* test: PageControllerTest DisPlayName 추가

* refactor: 접근 제어자 추가

* refactor: pathVariable에서 queryString 방식으로 변경

* refactor: 불필요한 sout 삭제

* refactor: resolveArgument에서 id값 반환하도록 변경

* refactor: DaoTest SpringBootTest에서 JdbcTest로 변경

* refactor: 테스트 격리 위한 코드 변경, SQL 쿼리 추가

* test: cartControllerTest 추가

* test: BasicAuthorizationExtractorTest 추가

* style: 코드 컨벤션에 따른 스타일 수정

+++ src/main/java/cart/service/AdminService.java
@@ -3,25 +3,23 @@ package cart.service;
 import cart.dao.ProductDao;
 import cart.dto.ProductRequestDto;
 import cart.entity.ProductEntity;
-import org.springframework.beans.factory.annotation.Autowired;
+import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 @Transactional
 public class AdminService {
     private final ProductDao productDao;
 
-    @Autowired
     public AdminService(ProductDao productDao) {
         this.productDao = productDao;
     }
 
 
     public int addProduct(ProductRequestDto productRequestDto) {
-        ProductEntity productEntity = new ProductEntity(productRequestDto.getName(), productRequestDto.getPrice(), productRequestDto.getImage());
+        ProductEntity productEntity = new ProductEntity(productRequestDto.getName(), productRequestDto.getPrice(),
+                productRequestDto.getImage());
         return productDao.insertProduct(productEntity);
     }
 
@@ -31,15 +29,25 @@ public class AdminService {
     }
 
     public void updateProduct(ProductRequestDto productRequestDto, int productId) {
-        ProductEntity productEntity = new ProductEntity(productId, productRequestDto.getName(), productRequestDto.getPrice(), productRequestDto.getImage());
+        if (!productDao.isProductExist(productId)) {
+            throw new IllegalArgumentException("수정하려는 제품이 존재하지 않습니다.");
+        }
+
+        ProductEntity productEntity = new ProductEntity(productId, productRequestDto.getName(),
+                productRequestDto.getPrice(), productRequestDto.getImage());
         productDao.updateProduct(productEntity);
     }
 
     public void deleteProduct(int productId) {
+        if (!productDao.isProductExist(productId)) {
+            throw new IllegalArgumentException("삭제하려는 제품이 존재하지 않습니다.");
+        }
+
         productDao.deleteProduct(productId);
     }
 
     public void deleteAll() {
         productDao.deleteAllProduct();
     }
+
 }
